Nadezda:Delete Kotlin
