--- src/student_nadezda_travkina/HelloWorld.java
@@ -1,48 +1,0 @@
-package student_nadezda_travkina;
-
-import java.util.Scanner;
-
-public class HelloWorld{
-    public static void main(String args[]){
-            System.out.println("Hello World...");
-    }
-}
-  public class Main {
-
-        public static void main(String[] args) {
-            Scanner num = new Scanner(System.in);
-            int first, second, result;
-            System.out.print("Enter first num: ");
-            first = num.nextInt();
-            System.out.print("Enter second num: ");
-            second = num.nextInt();
-            result = first + second;
-            System.out.println("Result is -" + result);
-        }
-  }
-
-  public class Variables {
-
-    oublic static void main(String[] args) {
-
-        int age = 20;
-        double varOne = 20.34;
-
-        boolean hungry = false; //false
-
-          char symbol = '4';
-
-          int variableA = 10;
-          int variableB = 220;
-          int result = variableA + variableB
-
-
-          System.out.println("variableA + variableB");
-
-
-    }
-}
-
-
-
-
