--- src/student_nadezda_travkina/Lesson2_task1.kt
@@ -1,14 +1,0 @@
-package student_nadezda_travkina
-
-class Lesson2_task1 {
-    public static void main(String[] args)
-    {Scanner num = new Scaner (System.in);
-        int first, second, result;
-        System.out.print("Enter first num: ");
-        first = num.Next.Int();
-        System.out.print("Enter second num: ");
-        second = num.Next.Int();
-        result = first + second;// чтобы ввести код для вычитания, деления и других действий, нужно заново задавать код для переменных? Если нет, то как программа понимает, какое действие выполнять?
-        System.out.print.ln("Result is -" + result)
-            }
-}
