--- src/student_nadezda_travkina/Lesson2_task2.kt
@@ -1,13 +1,0 @@
-package student_nadezda_travkina
-
-public class Lesson2_task2 {
-    @kotlin.jvm.JvmStatic
-    fun main(args: Array<String>) {
-        val `in` = Scanner(java.lang.System.`in`)
-        print("Input 1st number: ")
-        `in`.nextDouble()
-        print("Input 2nd number: ")
-        `in`.nextDouble()
-        println()
-    }
-}
