--- src/student_nadezda_travkina/Lesson2_task3.kt
@@ -1,4 +1,0 @@
-package student_nadezda_travkina
-
-public class Lesson2_task3 {
-}//Lesson2_task3 = Book
