--- src/student_nadezda_travkina/Lesson2_task5.kt
@@ -1,6 +1,0 @@
-package student_nadezda_travkina
-
-public class Lesson2_task5 { //Lesson2_task5 =BankAccount
-    public static void(String[] args {
-    }
-    }
