+++ src/student_nadezda_travkina/lesson1/VariablesExample.java
@@ -1,0 +1,23 @@
+package student_nadezda_travkina.lesson1;
+
+public class VariablesExample {
+
+    public static void main(String[] args) {
+
+        int age = 20;
+        double varOne = 20.34;
+
+        boolean hungry = false; //false
+
+        char symbol = '4';
+
+        int variableA = 10;
+        int variableB = 220;
+        int result = variableA + variableB;
+
+
+        System.out.println("variableA + variableB");
+
+
+    }
+}
